requirements.txt
asgiref==3.3.1
Django==3.1.7
django-phonenumber-field==5.0.0
msedge-selenium-tools==3.141.3
phonenumbers==8.12.19
pytz==2021.1
selenium==3.141.0
sqlparse==0.4.1
urllib3==1.26.3
